Extract TUI shell and session browser modules
